[balloon]: Cope with the unresponsive host
When the balloon size change is requested, the driver is filling/leaking the balloon in steps of 512 pages. The host is informed after every step and the driver always watis for an ACK. However, if none is delivered within 1000 ms timeout (iva an interrupt), the driver stops performing the operation. It would be better to finish the operation even in case the timeout expires.

Signed-Off-By: Martin Drab <martin.drab@virtuozzo.com>