netkvm: move RX limits calculation to separate procedure
This is preparation commit for lazy RX memory allocation.
Depending on number of RX buffer we calculate the kick limit
and low rx buffers limit. In future the number of RX buffers
become variable and will be incremented during lazy memory
allocation, so we will need to recalculate it at least several
times, so move it to separate procedure.

Signed-off-by: Yuri Benditovich <yuri.benditovich@daynix.com>